backend/requirements.txt
fastapi
uvicorn
openai
httpx
pyyaml
python-dotenv
pydantic
psutil
